ENH: don't need "double" distance map - only wastes space
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@3368 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Unstable/cmtkLabelCombinationLocalShapeBasedAveraging.cxx
@@ -50,7 +50,7 @@ cmtk::LabelCombinationLocalShapeBasedAveraging::AddAtlas( const UniformVolume::S
 {
   Superclass::AddAtlas( image, atlas );
 
-  this->m_AtlasDMaps.push_back( UniformDistanceMap<double>( *atlas, UniformDistanceMap<double>::SIGNED | UniformDistanceMap<double>::SQUARED ).Get() );
+  this->m_AtlasDMaps.push_back( UniformDistanceMap<float>( *atlas, UniformDistanceMap<float>::SIGNED | UniformDistanceMap<float>::SQUARED ).Get() );
 }
 
 cmtk::TypedArray::SmartPtr
